Changing pynapple core; adding abstract Base class and abstract BaseTsd class
