+++ pynapple/__init__.py
@@ -1,4 +1,4 @@
 __version__ = "0.5.1"
-from .core import *
+from .core import IntervalSet, Ts, Tsd, TsdFrame, TsdTensor, TsGroup, TsIndex, config
 from .io import *
 from .process import *
