+++ pynapple/core/config.py
@@ -1,4 +1,4 @@
-"""Package configurations.
+"""This module deals with package configurations. For now it includes only warning configurations.
 """
 
 
@@ -9,12 +9,40 @@ class PynappleConfig:
     This class includes all configuration settings that control the behavior of
     pynapple. It offers a structured way to access and modify settings.
 
-    Attributes:
-    -----------
-    suppress_conversion_warnings (bool):
+    Examples
+    --------
+    >>> import pynapple as nap
+    >>> import jax.numpy as jnp
+    >>> t = jnp.arange(3)
+    >>> print(t)
+    Array([0, 1, 2], dtype=int32)
+
+    >>> # Suppress warnings when converting a non-numpy array to numpy array
+    >>> nap.config.nap_config.suppress_conversion_warnings = True
+    >>> nap.Ts(t=t)
+    Time (s)
+    0.0
+    1.0
+    2.0
+    shape: 3
+
+    >>> # Restore to defaults
+    >>> nap.config.nap_config.restore_defaults()
+    >>> nap.Ts(t=t)
+    /mnt/home/gviejo/pynapple/pynapple/core/time_series.py:151: UserWarning: Converting 't' to n
+    umpy.array. The provided array was of type 'ArrayImpl'.
+      warnings.warn(
+    Time (s)
+    0.0
+    1.0
+    2.0
+    shape: 3
+
+    Attributes
+    ----------
+    suppress_conversion_warnings : boolean
         Determines whether to suppress warnings when automatically converting non-NumPy
-        array-like objects to NumPy arrays.
-        This is useful for users who frequently work with array-like objects from other
+        array-like objects to NumPy arrays. This is useful for users who frequently work with array-like objects from other
         libraries (e.g., JAX, TensorFlow) and prefer not to receive warnings for automatic
         conversions. Defaults to False, which means warnings will be shown.
     """
@@ -25,7 +53,7 @@ class PynappleConfig:
     @property
     def suppress_conversion_warnings(self):
         """
-        bool: Gets or sets the suppression state for conversion warnings. When set to True,
+        Gets or sets the suppression state for conversion warnings. When set to True,
         warnings for automatic conversions of non-NumPy array-like objects to NumPy arrays
         are suppressed. Ensures that only boolean values are assigned.
         """
