+++ tests/__init__.py
@@ -1,2 +1,1 @@
 """Unit test package for pynapple."""
-from . import mock
