+++ tests/test_time_series.py
@@ -2,7 +2,7 @@
 # @Author: gviejo
 # @Date:   2022-04-01 09:57:55
 # @Last Modified by:   Guillaume Viejo
-# @Last Modified time: 2023-12-07 14:01:09
+# @Last Modified time: 2024-02-13 17:36:43
 #!/usr/bin/env python
 
 """Tests of time series for `pynapple` package."""
@@ -11,8 +11,6 @@ import pynapple as nap
 import numpy as np
 import pandas as pd
 import pytest
-
-from pynapple.core.time_series import TsdTensor
 
 
 def test_create_tsd():
@@ -23,6 +21,10 @@ def test_create_empty_tsd():
     tsd = nap.Tsd(t=np.array([]), d=np.array([]))
     assert len(tsd) == 0
 
+@pytest.mark.filterwarnings("ignore")
+def test_create_tsd_from_number():
+    tsd = nap.Tsd(t=1, d=2)
+
 def test_create_tsdframe():
     tsdframe = nap.TsdFrame(t=np.arange(100), d=np.random.rand(100, 4))
     assert isinstance(tsdframe, nap.TsdFrame)
@@ -31,12 +33,13 @@ def test_create_tsdframe():
     assert isinstance(tsdframe, nap.TsdFrame)
     assert np.all(tsdframe.columns == np.array(['a', 'b', 'c', 'd']))
 
+@pytest.mark.filterwarnings("ignore")
 def test_create_empty_tsdframe():
-    tsdframe = nap.TsdFrame(t=np.array([]), d=np.array([]))
+    tsdframe = nap.TsdFrame(t=np.array([]), d=np.empty(shape=(0, 2)))
     assert len(tsdframe) == 0
     assert isinstance(tsdframe, nap.TsdFrame)
 
-    with pytest.raises(RuntimeError):
+    with pytest.raises(AssertionError):
         tsdframe = nap.TsdFrame(t=np.arange(100))
 
 def test_create_1d_tsdframe():
@@ -185,6 +188,7 @@ def test_concatenate_tsd():
 
     np.testing.assert_array_equal(new_tsd.values, tsd.values)
 
+@pytest.mark.filterwarnings("ignore")
 def test_create_tsdtensor():
     tsd = nap.TsdTensor(t=np.arange(100), d=np.random.rand(100, 3, 2))
     assert isinstance(tsd, nap.TsdTensor)
@@ -193,17 +197,18 @@ def test_create_tsdtensor():
     assert isinstance(tsd, nap.TsdTensor)    
 
 def test_create_empty_tsd():
-    tsd = nap.TsdTensor(t=np.array([]), d=np.array([[[]]]).T)
+    tsd = nap.TsdTensor(t=np.array([]), d=np.empty(shape=(0,2,3)))
     assert len(tsd) == 0
 
+@pytest.mark.filterwarnings("ignore")
 def test_raise_error_tsdtensor_init():
-    with pytest.raises(RuntimeError, match=r"Missing argument d when initializing TsdTensor"):
+    with pytest.raises(RuntimeError, match=r"Unknown format for d. Accepted formats are numpy.ndarray, list, tuple or any array-like objects."):
         nap.TsdTensor(t=np.arange(100), d=None)
 
-    with pytest.raises(AssertionError, match=r"Data should have more than 2 dimensions. If ndim < 3, use TsdFrame or Tsd object"):
-        nap.TsdTensor(t=np.arange(100), d=np.random.rand(100, 10))
+    # with pytest.raises(AssertionError, match=r"Data should have more than 2 dimensions. If ndim < 3, use TsdFrame or Tsd object"):
+    #     nap.TsdTensor(t=np.arange(100), d=np.random.rand(100, 10))
 
-    with pytest.raises(ValueError):#, match=r"Length of values (10) does not match length of index (100)"):
+    with pytest.raises(AssertionError):#, match=r"Length of values (10) does not match length of index (100)"):
         nap.TsdTensor(t=np.arange(100), d=np.random.rand(10, 10,3))
 
 def test_index_error():
@@ -251,12 +256,15 @@ def test_properties():
     with pytest.raises(RuntimeError):
         tsd.rate = 0
 
-def test_abstract_class():
-    class DummyTsd(nap.core.time_series._AbstractTsd):
-        def __init__(self):
-            super().__init__()
+def test_base_tsd_class():
+    class DummyTsd(nap.core.time_series.BaseTsd):
+        def __init__(self, t, d):
+            super().__init__(t, d)
 
-    tsd = DummyTsd()
+        def __getitem__(self, key):
+            return self.values.__getitem__(key)
+
+    tsd = DummyTsd([], [])
     assert np.isnan(tsd.rate)
     assert isinstance(tsd.index, nap.TsIndex)
     assert isinstance(tsd.values, np.ndarray)
@@ -332,7 +340,7 @@ class Test_Time_Series_1:
         np.testing.assert_array_almost_equal(tsdframe.values[::10], tsdframe2.values)
 
     def test_value_from_value_error(self, tsd):
-        with pytest.raises(RuntimeError):
+        with pytest.raises(AssertionError, match=r"First argument should be an instance of Tsd, TsdFrame or TsdTensor"):
             tsd.value_from(np.arange(10))
         
     def test_value_from_with_restrict(self, tsd):
@@ -402,7 +410,7 @@ class Test_Time_Series_1:
             np.testing.assert_array_equal(tsd.index.values, new_tsd.index.values)
             np.testing.assert_array_equal(tsd.values, new_tsd.values)
 
-            tsd.values[tsd.values>0.9] = np.NaN         
+            tsd.values[tsd.values>0.9] = np.NaN
             new_tsd = tsd.dropna()
             assert not np.all(np.isnan(new_tsd))
             tokeep = np.array([~np.any(np.isnan(tsd[i])) for i in range(len(tsd))])            
@@ -721,9 +729,9 @@ class Test_Time_Series_2:
         tsd2 = tsd.interpolate(ts)
         np.testing.assert_array_almost_equal(tsd2.values, y)
         
-        with pytest.raises(RuntimeError) as e:
+        with pytest.raises(AssertionError) as e:
             tsd.interpolate([0, 1, 2])
-        assert str(e.value) == "First argument should be an instance of Ts, Tsd or TsdFrame"
+        assert str(e.value) == "First argument should be an instance of Ts, Tsd, TsdFrame or TsdTensor"
 
         # Right left
         ep = nap.IntervalSet(start=0, end=5)
@@ -953,9 +961,9 @@ class Test_Time_Series_3:
         tsdframe2 = tsdframe.interpolate(ts)
         np.testing.assert_array_almost_equal(tsdframe2.values, data_stack)
         
-        with pytest.raises(RuntimeError) as e:
+        with pytest.raises(AssertionError) as e:
             tsdframe.interpolate([0, 1, 2])
-        assert str(e.value) == "First argument should be an instance of Ts, Tsd or TsdFrame"
+        assert str(e.value) == "First argument should be an instance of Ts, Tsd, TsdFrame or TsdTensor"
 
         # Right left
         ep = nap.IntervalSet(start=0, end=5)
@@ -1253,9 +1261,9 @@ class Test_Time_Series_5:
         tsdtensor2 = tsdtensor.interpolate(ts)
         np.testing.assert_array_almost_equal(tsdtensor2.values, data_stack)
         
-        with pytest.raises(RuntimeError) as e:
+        with pytest.raises(AssertionError) as e:
             tsdtensor.interpolate([0, 1, 2])
-        assert str(e.value) == "First argument should be an instance of Ts, Tsd or TsdFrame"
+        assert str(e.value) == "First argument should be an instance of Ts, Tsd, TsdFrame or TsdTensor"
 
         # Right left
         ep = nap.IntervalSet(start=0, end=5)
